requirements.txt
absl-py==1.1.0
appdirs==1.4.4
attrs==21.4.0
autopep8==1.6.0
black==22.6.0
certifi==2022.6.15
click==8.1.3
cycler==0.11.0
flake8==4.0.1
fonttools==4.34.4
future==0.18.2
importlib-metadata==4.12.0
iso8601==1.0.2
kiwisolver==1.4.3
matplotlib==3.5.2
mccabe==0.6.1
mediapipe==0.8.10.1
mypy-extensions==0.4.3
numpy==1.23.1
opencv-contrib-python==4.6.0.66
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
protobuf==3.20.1
pycodestyle==2.8.0
pyee==8.2.2
pyflakes==2.4.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyserial==3.5
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
urllib3==1.26.10
websockets==10.3
zipp==3.8.0
websocket-client
Pillow
requests
requests-toolbelt